Fixed XmlTransferMapUtils.
